requirements.txt
# AWS CDK and infrastructure dependencies
aws-cdk-lib==2.139.0
constructs==10.3.0
aws-cdk-lib[core]==2.139.0

# AWS and AWS utilities
boto3==1.43.2
botocore==1.43.2

# Phase 2: Content Ingestion Pipeline
anthropic>=0.45.0
apify-client==1.7.0
pydantic>=2.8.0
requests==2.31.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0

# Code quality
black==24.1.1
flake8==7.0.0
pylint==3.0.3

# Type checking
mypy==1.8.0
aws-lambda-typing>=2.20.0

# Development utilities
python-dotenv==1.0.0
